fix(config): remove system config source
Refs #5148
