+++ src/core/config/index.js
@@ -5,7 +5,6 @@ export { default as inlinePluginOptionsFactorization } from "./factorization/inl
 export { default as storeOptionsFactorization } from "./factorization/store"
 export { default as optionsFromQuery } from "./sources/query"
 export { default as optionsFromURL } from "./sources/url"
-export { default as optionsFromSystem } from "./sources/system"
 export { default as optionsFromRuntime } from "./sources/runtime"
 export { default as defaultOptions } from "./defaults"
 export { default as mergeOptions } from "./merge"
