+++ src/core/plugins/configs/actions.js
@@ -1,3 +1,8 @@
+/**
+ * @prettier
+ */
+import { parseYamlConfig } from "./helpers"
+
 export const UPDATE_CONFIGS = "configs_update"
 export const TOGGLE_CONFIGS = "configs_toggle"
 
@@ -6,7 +11,7 @@ export function update(configName, configValue) {
   return {
     type: UPDATE_CONFIGS,
     payload: {
-      [configName]: configValue
+      [configName]: configValue,
     },
   }
 }
@@ -19,8 +24,35 @@ export function toggle(configName) {
   }
 }
 
-
 // Hook
 export const loaded = () => () => {
   // noop
 }
+
+export const downloadConfig = (req) => (system) => {
+  const {
+    fn: { fetch },
+  } = system
+
+  return fetch(req)
+}
+
+export const getConfigByUrl =
+  (req, cb) =>
+  ({ specActions, configsActions }) => {
+    if (req) {
+      return configsActions.downloadConfig(req).then(next, next)
+    }
+
+    function next(res) {
+      if (res instanceof Error || res.status >= 400) {
+        specActions.updateLoadingStatus("failedConfig")
+        specActions.updateLoadingStatus("failedConfig")
+        specActions.updateUrl("")
+        console.error(res.statusText + " " + req.url)
+        cb(null)
+      } else {
+        cb(parseYamlConfig(res.text))
+      }
+    }
+  }
