--- swagger-config.yaml
@@ -1,4 +1,0 @@
----
-url: "https://petstore.swagger.io/v2/swagger.json"
-dom_id: "#swagger-ui"
-validatorUrl: "https://validator.swagger.io/validator"
